examples/python_client/requirements.txt
sdk-4mica==0.4.2
httpx>=0.27.2
python-dotenv>=1.0.1
